org per-entrypoint CRUD funcs together; all fixtures in conftest
