+++ tests/test_database.py
@@ -13,20 +13,14 @@ from sqlmodel import Session, SQLModel
 
 import pangeo_forge_orchestrator.abstractions as abstractions
 
-from .conftest import (
-    CreateFixtures,
-    DeleteFixtures,
-    ModelFixture,
-    ReadFixtures,
-    UpdateFixtures,
+from .conftest import CreateFixtures, DeleteFixtures, ModelFixture, ReadFixtures, UpdateFixtures
+from .entrypoints import (
     _IntTypeError,
     _MissingFieldError,
     _NonexistentTableError,
     _StrTypeError,
     clear_table,
 )
-
-ENTRYPOINTS = ["db", "abstract-funcs", "client", "cli"]
 
 
 def commit_to_session(session: Session, model: SQLModel):
@@ -112,11 +106,6 @@ def test_registration(session, models_with_kwargs):
 class TestCreate(CreateFixtures):
     """Container for tests of database entry creation and associated failure modes"""
 
-    @pytest.fixture
-    def model_to_create(self, models_with_kwargs):
-        kw_0, _, _ = models_with_kwargs.kwargs
-        return models_with_kwargs.models, kw_0.request, kw_0.blank_opts
-
     @staticmethod
     def get_error(func: Callable, failure_mode: str):
         entrypoint = get_entrypoint(func)
@@ -190,21 +179,6 @@ class TestCreate(CreateFixtures):
 class TestRead(ReadFixtures):
     """Container for tests of reading from database"""
 
-    @pytest.fixture
-    def models_to_read(self, models_with_kwargs: ModelFixture):
-        models = models_with_kwargs.models
-        kw_0, kw_1, _ = models_with_kwargs.kwargs
-        model_0 = models.table(**kw_0.request)
-        model_1 = models.table(**kw_1.request)
-        return models, (model_0, model_1)
-
-    @pytest.fixture
-    def single_model_to_read(self, models_with_kwargs: ModelFixture):
-        models = models_with_kwargs.models
-        _, _, kw_2 = models_with_kwargs.kwargs
-        table = models.table(**kw_2.request)
-        return models, table
-
     @staticmethod
     def get_error(func: Callable):
         entrypoint = get_entrypoint(func)
@@ -272,16 +246,6 @@ class TestRead(ReadFixtures):
 class TestUpdate(UpdateFixtures):
     """Container for tests of updating existing entries in database"""
 
-    @pytest.fixture(scope="session")
-    def model_to_update(self, models_with_kwargs):
-        models = models_with_kwargs.models
-        kw_0, kw_1, _ = models_with_kwargs.kwargs
-        table = models.table(**kw_0.request)
-        different_kws = copy.deepcopy(kw_1.request)
-        key = list(different_kws)[0]
-        update_with = {key: different_kws.pop(key)}
-        return models, table, update_with
-
     @staticmethod
     def get_error(func: Callable):
         entrypoint = get_entrypoint(func)
@@ -329,13 +293,6 @@ class TestUpdate(UpdateFixtures):
 
 class TestDelete(DeleteFixtures):
     """Container for tests of deleting existing entries in database"""
-
-    @pytest.fixture
-    def model_to_delete(self, models_with_kwargs):
-        models = models_with_kwargs.models
-        kw_0, _, _ = models_with_kwargs.kwargs
-        table = models.table(**kw_0.request)
-        return models, table
 
     @staticmethod
     def get_error(func: Callable):
